+++ src/Users/Tables/UserIndex.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Enraiged\Users\Tables\Builders;
+namespace Enraiged\Users\Tables;
 
 use Enraiged\Users\Models\User;
 use Enraiged\Users\Tables\Exporters\IndexExporter;
@@ -22,19 +22,7 @@ class UserIndex extends TableBuilder implements ProvidesDefaultSort, ProvidesTab
     protected $resource = IndexResource::class;
 
     /** @var  string  The template json file path. */
-    protected string $template = __DIR__.'/../Templates/user-index.json';
-
-    /**
-     *  Determine whether or not a provided object is secure.
-     *
-     *  @param  array|object  $object
-     *  @param  \Illuminate\Database\Eloquent\Model|null  $model = null
-     *  @return bool
-     *
-    protected function assertSecure($object, $model = null): bool
-    {
-        return true;
-    }*/
+    protected string $template = __DIR__.'/Templates/user-index.json';
 
     /**
      *  Apply default sort criteria to this table builder.
@@ -53,12 +41,24 @@ class UserIndex extends TableBuilder implements ProvidesDefaultSort, ProvidesTab
      *
      *  @param bool $param
      */
-    public function filterActive($param)
+    public function filterStatus($param)
     {
-        $active = filter_var($param, FILTER_VALIDATE_BOOLEAN);
-
-        $this->builder
-            ->where('users.is_active', $active);
+        switch ($param) {
+            case 'active':
+                $this->builder
+                    ->where('users.is_active', true)
+                    ->whereNull('users.deleted_at');
+                break;
+            case 'inactive':
+                $this->builder
+                    ->where('users.is_active', false)
+                    ->whereNull('users.deleted_at');
+                break;
+            case 'deleted':
+                $this->builder
+                    ->whereNotNull('users.deleted_at');
+                break;
+        }
     }
 
     /**
@@ -71,15 +71,11 @@ class UserIndex extends TableBuilder implements ProvidesDefaultSort, ProvidesTab
         $builder = $this->model::withTrashed()
             ->select('users.*')
             ->join('profiles', 'profiles.id', '=', 'users.profile_id')
+            ->leftJoin('addresses', fn ($join)
+                => $join->on('addresses.addressable_id', '=', 'profiles.id')
+                    ->where('addresses.addressable_type', '=', 'profile'))
+            ->leftJoin('countries', 'countries.id', '=', 'addresses.country_id')
             ->where('users.is_hidden', false);
-
-        if (!$this->request()->hasFilter('active')) {
-            $builder->where('users.is_active', true);
-        }
-
-        if (!$this->request()->hasFilter('deleted')) {
-            $builder->whereNull('users.deleted_at');
-        }
 
         return $builder;
     }